requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
beautifulsoup4
requests
